refactor(replication): move queue contracts into crate
